Pulled out userAdapter to a central database adapter
