+++ src/databaseAdapters/user.ts
@@ -2,25 +2,7 @@ import User from '../models/user';
 import { IUser } from '../gameplay/types';
 import { S3AvatarSet } from '../clients/s3/S3Agent';
 
-interface DatabaseAdapter {
-  getUser(username: string): Promise<IUser>;
-  setAvatarLinks(
-    username: string,
-    resLink: string,
-    spyLink: string,
-  ): Promise<void>;
-  muteUser(userCallingMute: IUser, usernameToMute: string): Promise<void>;
-  unmuteUser(userCallingUnmute: IUser, usernameToUnmute: string): Promise<void>;
-  resetAvatar(username: string): Promise<void>;
-  setAvatarAndUpdateLibrary(
-    username: string,
-    avatarSet: S3AvatarSet,
-    librarySize: number,
-  ): Promise<void>;
-  removeAvatar(username: string, avatarSet: S3AvatarSet): Promise<void>;
-}
-
-class MongoUserAdapter implements DatabaseAdapter {
+export class MongoUserAdapter {
   async getUser(username: string): Promise<IUser> {
     return (await User.findOne({
       usernameLower: username.toLowerCase(),
@@ -95,6 +77,3 @@ class MongoUserAdapter implements DatabaseAdapter {
     await user.save();
   }
 }
-
-const userAdapter = new MongoUserAdapter();
-export default userAdapter;
